+++ components/core/Loader.vue
@@ -1,5 +1,5 @@
-<template>
-  <div class="loader-container fixed" v-if="isVisible">
+<template functional>
+  <div class="loader-container fixed">
     <div class="loader-inner-container fixed">
       <div class="spinner relative">
         <div class="double-bounce1 absolute w-100 brdr-circle bg-cl-th-success" />
@@ -7,9 +7,9 @@
       </div>
       <div
         class="loader-message-container mt15 py5 px15 align-center h6 cl-white"
-        v-if="message"
+        v-if="props.message"
       >
-        {{ message }}
+        {{ props.message }}
       </div>
     </div>
   </div>
@@ -20,30 +20,11 @@ import { mapState } from 'vuex'
 
 export default {
   name: 'Loader',
-  data () {
-    return {
-      message: null
+  props: {
+    message: {
+      type: String | null,
+      required: true
     }
-  },
-  methods: {
-    show (message = null) {
-      this.message = message
-      this.$store.commit('ui/setLoader', true)
-    },
-    hide () {
-      this.$store.commit('ui/setLoader', false)
-    }
-  },
-  computed: mapState({
-    isVisible: state => state.ui.loader
-  }),
-  beforeMount () {
-    this.$bus.$on('notification-progress-start', this.show)
-    this.$bus.$on('notification-progress-stop', this.hide)
-  },
-  beforeDestroy () {
-    this.$bus.$off('notification-progress-start', this.show)
-    this.$bus.$off('notification-progress-stop', this.hide)
   }
 }
 </script>