+++ components/core/blocks/MainSlider/HeadImage.vue
@@ -1,42 +1,36 @@
-<template>
+<template functional>
   <section class="head-image w-100 bg-cl-th-accent cl-white">
-    <div class="container w-100 h-100 cl-black" v-lazy:background-image="currentImage.image" v-if="currentImage">
+    <div
+      class="head-image__container container w-100 h-100 cl-black"
+      :class="{'webp': props.supportsWebp}"
+      v-if="props.currentImage"
+    >
       <div class="head-image-content">
         <h1 class="title" data-testid="mainSliderTitle">
-          {{ currentImage.title }}
+          {{ props.currentImage.title }}
         </h1>
         <p
           class="subtitle mb0 serif h3"
           data-testid="mainSliderSubtitle"
         >
-          {{ currentImage.subtitle }}
+          {{ props.currentImage.subtitle }}
         </p>
-        <!-- <div class="align-center inline-flex">
-          <button-outline :link="currentImage.link" color="light">
-            {{ currentImage.button_text }}
-          </button-outline>
-        </div> -->
       </div>
     </div>
   </section>
 </template>
 
 <script>
-// import ButtonOutline from 'theme/components/theme/ButtonOutline'
-import { mapGetters } from 'vuex'
-
 export default {
-  components: {
-    // ButtonOutline
-  },
-  data () {
-    return {
+  props: {
+    currentImage: {
+      type: Object,
+      required: true
+    },
+    supportsWebp: {
+      type: Boolean,
+      required: true
     }
-  },
-  computed: {
-    ...mapGetters({
-      currentImage: 'promoted/getHeadImage'
-    })
   }
 }
 </script>
@@ -70,6 +64,50 @@ export default {
 
 .head-image {
   height: 640px;
+
+  &__container {
+    background-image: url('/assets/full_width_banner-800.jpg');
+    @media (min-width: 850px) {
+      background-image: url('/assets/full_width_banner-1000.jpg');
+    }
+    @media (min-width: 1050px) {
+      background-image: url('/assets/full_width_banner-1200.jpg');
+    }
+    @media (min-width: 1250px) {
+      background-image: url('/assets/full_width_banner-1400.jpg');
+    }
+    @media (min-width: 1450px) {
+      background-image: url('/assets/full_width_banner-1600.jpg');
+    }
+    @media (min-width: 1650px) {
+      background-image: url('/assets/full_width_banner-1800.jpg');
+    }
+    @media (min-width: 1850px) {
+      background-image: url('/assets/full_width_banner.jpg');
+    }
+
+    &.webp {
+      background-image: url('/assets/full_width_banner-800.webp');
+      @media (min-width: 850px) {
+        background-image: url('/assets/full_width_banner-1000.webp');
+      }
+      @media (min-width: 1050px) {
+        background-image: url('/assets/full_width_banner-1200.webp');
+      }
+      @media (min-width: 1250px) {
+        background-image: url('/assets/full_width_banner-1400.webp');
+      }
+      @media (min-width: 1450px) {
+        background-image: url('/assets/full_width_banner-1600.webp');
+      }
+      @media (min-width: 1650px) {
+        background-image: url('/assets/full_width_banner-1800.webp');
+      }
+      @media (min-width: 1850px) {
+        background-image: url('/assets/full_width_banner.webp');
+      }
+    }
+  }
 }
 .container {
   background-size: cover;