+++ components/core/blocks/SearchPanel/SearchPanel.vue
@@ -2,6 +2,8 @@
   <div
     class="searchpanel fixed mw-100 bg-cl-primary cl-accent"
     data-testid="searchPanel"
+    ref="searchPanel"
+    :style="{ 'max-height': `${windowHelper}px` }"
   >
     <div class="close-icon-row">
       <i
@@ -79,15 +81,15 @@
 
 <script>
 import SearchPanel from '@vue-storefront/core/compatibility/components/blocks/SearchPanel/SearchPanel'
-import ProductTile from 'theme/components/core/ProductTile'
 import VueOfflineMixin from 'vue-offline/mixin'
 import CategoryPanel from 'theme/components/core/blocks/Category/CategoryPanel'
 import { minLength } from 'vuelidate/lib/validators'
 import { disableBodyScroll, clearAllBodyScrollLocks } from 'body-scroll-lock'
+import { windowHelper } from 'theme/helpers'
 
 export default {
   components: {
-    ProductTile,
+    ProductTile: () => import(/* webpackChunkName: "vsf-product-tile" */ 'theme/components/core/ProductTile.vue'),
     CategoryPanel
   },
   mixins: [SearchPanel, VueOfflineMixin],
@@ -131,20 +133,26 @@ export default {
         msg = 'No results were found.'
       }
       return msg
+    },
+    windowHelper () {
+      return windowHelper.height
     }
   },
   watch: {
     categories () {
       this.selectedCategoryIds = []
+    },
+    isOpen (state) {
+      if (state) {
+        this.$nextTick(() => {
+          disableBodyScroll(this.$refs.searchPanel)
+        })
+      } else clearAllBodyScrollLocks()
     }
   },
   mounted () {
     // add autofocus to search input field
     this.$refs.search.focus()
-    disableBodyScroll(this.$el)
-  },
-  destroyed () {
-    clearAllBodyScrollLocks()
   }
 }
 </script>
@@ -156,6 +164,7 @@ export default {
 
 .searchpanel {
   height: 100vh;
+  min-height: 100%;
   width: 800px;
   top: 0;
   right: 0;