+++ components/theme/blocks/PromotedOffers/PromotedOffers.vue
@@ -2,14 +2,14 @@
   <section v-if="!singleBanner" class="offers container my30 px15 cl-black">
     <div class="row">
       <div
-        class="offer-container col-xs-12 col-sm-6 pb15"
+        class="offer-container offer-container--main col-xs-12 col-sm-6 pb15"
         v-for="(banner, index) in banners.mainBanners"
         :key="index"
       >
         <router-link :to="localizedRoute(banner.link)">
           <div
-            class="offer"
-            v-lazy:background-image="banner.image"
+            class="offer offer--main"
+            :class="{'webp': supportsWebp}"
           >
             <h2 class="title m0 h1">
               {{ banner.title }}
@@ -27,19 +27,25 @@
           v-for="(banner, index) in banners.smallBanners"
           :key="index"
         >
-          <router-link :to="localizedRoute(banner.link)">
-            <div
-              class="offer offer-small border-box p5 flex bg-cl-th-accent"
-              v-lazy:background-image="banner.image"
-            >
-              <h2 class="title m0 h1">
-                {{ banner.title }}
-              </h2>
-              <p class="subtitle m0 serif h3 uppercase">
-                {{ banner.subtitle }}
-              </p>
-            </div>
-          </router-link>
+          <lazy-hydrate when-visible>
+            <router-link :to="localizedRoute(banner.link)">
+              <div
+                class="offer offer-small border-box p5 flex bg-cl-th-accent"
+                :class="{
+                  'webp': supportsWebp, 
+                  'offer-shine-on': index === 0,
+                  'offer-spring': index === 1
+                  }"
+              >
+                <h2 class="title m0 h1">
+                  {{ banner.title }}
+                </h2>
+                <p class="subtitle m0 serif h3 uppercase">
+                  {{ banner.subtitle }}
+                </p>
+              </div>
+            </router-link>
+          </lazy-hydrate>
         </div>
       </div>
     </div>
@@ -51,19 +57,20 @@
         v-for="(banner, index) in banners.productBanners"
         :key="index"
       >
-        <router-link :to="localizedRoute(banner.link)">
-          <div
-            class="offer offer-product border-box p5 flex bg-cl-th-accent"
-            v-lazy:background-image="banner.image"
-          >
-            <h2 class="title m0 h1">
-              {{ banner.title }}
-            </h2>
-            <p class="subtitle m0 serif h3 uppercase">
-              {{ banner.subtitle }}
-            </p>
-          </div>
-        </router-link>
+        <lazy-hydrate when-visible>
+          <router-link :to="localizedRoute(banner.link)">
+            <div
+              class="offer offer-spring offer-product border-box p5 flex bg-cl-th-accent"
+            >
+              <h2 class="title m0 h1">
+                {{ banner.title }}
+              </h2>
+              <p class="subtitle m0 serif h3 uppercase">
+                {{ banner.subtitle }}
+              </p>
+            </div>
+          </router-link>
+        </lazy-hydrate>
       </div>
     </div>
   </section>
@@ -71,6 +78,95 @@
 
 <script>
 import { mapGetters, mapActions } from 'vuex'
+import LazyHydrate from 'vue-lazy-hydration'
+import { currentStoreView } from '@vue-storefront/core/lib/multistore'
+
+const offers = {
+  de: {
+    "mainBanners": [
+      {
+        "title": "Lässige Büro",
+        "subtitle": "Kollektion",
+        "link": "/women/frauen-20"
+      }
+    ],
+    "smallBanners": [
+      {
+        "title": "Glänzen Sie mit",
+        "subtitle": "Accessoires",
+        "link": "/men/herren-11"
+      },
+      {
+        "title": "Der Frühling kommt",
+        "subtitle": "Hüte",
+        "link": "/gear/gerat-3"
+      }
+    ],
+    "productBanners": [
+      {
+        "title": "Der Frühling kommt",
+        "subtitle": "Hüte",
+        "link": "/gear/gerat-3"
+      }
+    ]
+  },
+  it: {
+    "mainBanners": [
+      {
+        "title": "Ufficio casual",
+        "subtitle": "Collezione",
+        "link": "/women/la-donne-20"
+      }
+    ],
+    "smallBanners": [
+      {
+        "title": "Brilla",
+        "subtitle": "Accessori",
+        "link": "/men/signori-11"
+      },
+      {
+        "title": "La primavera sta arrivando",
+        "subtitle": "Cappelli",
+        "link": "/gear/equipaggiamento-3"
+      }
+    ],
+    "productBanners": [
+      {
+        "title": "La primavera sta arrivando",
+        "subtitle": "Cappelli",
+        "link": "/gear/equipaggiamento-3"
+      }
+    ]
+  },
+  default: {
+    "mainBanners": [
+      {
+        "title": "Office casual",
+        "subtitle": "Collection",
+        "link": "/women.html"
+      }
+    ],
+    "smallBanners": [
+      {
+        "title": "Shine on",
+        "subtitle": "Accessories",
+        "link": "/men.html"
+      },
+      {
+        "title": "Spring is coming",
+        "subtitle": "Hats",
+        "link": "/gear.html"
+      }
+    ],
+    "productBanners": [
+      {
+        "title": "Spring is coming",
+        "subtitle": "Hats",
+        "link": "/gear.html"
+      }
+    ]
+  }
+}
 
 export default {
   name: 'PromotedOffers',
@@ -81,18 +177,18 @@ export default {
       default: false
     }
   },
+  components: {
+    LazyHydrate
+  },
   computed: {
-    ...mapGetters({
-      banners: 'promoted/getPromotedOffers'
-    })
-  },
-  async created () {
-    await this.updatePromotedOffers()
-  },
-  methods: {
-    ...mapActions({
-      updatePromotedOffers: 'promoted/updatePromotedOffers'
-    })
+    banners () {
+      const { storeCode } = currentStoreView()
+      
+      return offers[storeCode] || offers.default
+    },
+    supportsWebp () {
+      return this.$store.state.ui.supportsWebp
+    }
   }
 }
 </script>
@@ -101,6 +197,11 @@ export default {
   .offer-container {
     &:last-child {
       padding-bottom: 0;
+    }
+    &--main {
+      @media (max-width: 576px) {
+        height: calc(100vh - 84px);
+      }
     }
   }
   .offer {
@@ -114,12 +215,135 @@ export default {
     opacity: 1;
     transition: 0.3s all;
 
-    &:hover {
-      opacity: 0.9;
-    }
-
     @media (max-width: 767px) {
       height: 200px;
+    }
+
+    &--main {
+        background-image: url('/assets/ban1-320.jpg');
+          @media (min-width: 375px) {
+            background-image: url('/assets/ban1-520.jpg');
+          }
+
+          @media (min-width: 576px) {
+            background-image: url('/assets/ban1-320.jpg');
+          }
+
+          @media (min-width: 768px) {
+            background-image: url('/assets/ban1-460.jpg');
+          }
+
+          @media (min-width: 1200px) {
+            background-image: url('/assets/ban1.jpg');
+          }
+
+        &.webp {
+          background-image: url('/assets/ban1-320.webp');
+          @media (min-width: 375px) {
+            background-image: url('/assets/ban1-520.webp');
+          }
+
+          @media (min-width: 576px) {
+            background-image: url('/assets/ban1-320.webp');
+          }
+
+          @media (min-width: 768px) {
+            background-image: url('/assets/ban1-460.webp');
+          }
+
+          @media (min-width: 1200px) {
+            background-image: url('/assets/ban1.webp');
+          }
+        }
+        @media (max-width: 575px) {
+          height: 100%;
+        }
+    }
+
+    &-shine-on {
+        background-image: url('/assets/ban2-320.jpg');
+          @media (min-width: 375px) {
+            background-image: url('/assets/ban2-520.jpg');
+          }
+
+          @media (min-width: 576px) {
+            background-image: url('/assets/ban2-320.jpg');
+          }
+
+          @media (min-width: 768px) {
+            background-image: url('/assets/ban2-460.jpg');
+          }
+
+          @media (min-width: 1200px) {
+            background-image: url('/assets/ban2.jpg');
+          }
+
+        &.webp {
+          background-image: url('/assets/ban2-320.webp');
+          @media (min-width: 375px) {
+            background-image: url('/assets/ban2-520.webp');
+          }
+
+          @media (min-width: 576px) {
+            background-image: url('/assets/ban2-320.webp');
+          }
+
+          @media (min-width: 768px) {
+            background-image: url('/assets/ban2-460.webp');
+          }
+
+          @media (min-width: 1200px) {
+            background-image: url('/assets/ban2.webp');
+          }
+        }
+        @media (max-width: 575px) {
+          height: 100%;
+        }
+    }
+
+    &-spring {
+        background-image: url('/assets/ban3-320.jpg');
+          @media (min-width: 375px) {
+            background-image: url('/assets/ban3-520.jpg');
+          }
+
+          @media (min-width: 576px) {
+            background-image: url('/assets/ban3-320.jpg');
+          }
+
+          @media (min-width: 768px) {
+            background-image: url('/assets/ban3-460.jpg');
+          }
+
+          @media (min-width: 1200px) {
+            background-image: url('/assets/ban3.jpg');
+          }
+
+        &.webp {
+          background-image: url('/assets/ban3-320.webp');
+          @media (min-width: 375px) {
+            background-image: url('/assets/ban3-520.webp');
+          }
+
+          @media (min-width: 576px) {
+            background-image: url('/assets/ban3-320.webp');
+          }
+
+          @media (min-width: 768px) {
+            background-image: url('/assets/ban3-460.webp');
+          }
+
+          @media (min-width: 1200px) {
+            background-image: url('/assets/ban3.webp');
+          }
+        }
+        @media (max-width: 575px) {
+          height: 100%;
+        }
+    }
+
+    &:hover {
+      opacity: 0.9;
     }
 
     .title {
@@ -150,10 +374,6 @@ export default {
   .offer-product {
     height: 330px;
     background-position: 50% 20%;
-
-    @media (max-width: 767px) {
-      height: 330px;
-    }
   }
   .title {
     @media (max-width: 767px) {